Funcionalidade de play da música + estilização

+++ modulo2/projeto-labefy/labefy/src/pages/AllPlaylists/AllPlaylists.js
@@ -2,8 +2,16 @@ import axios from "axios";
 import React from "react";
 import { axiosConfig } from "../../constants/headers";
 import { BASE_URL } from "../../constants/urls";
-import imgDelete from "./img/delete.png";
-import imgPlaylist from "./img/playlist.png"
+
+import imgCd from "../../img/cd.png";
+import imgDelete from "../../img/delete.png";
+
+import { MainContainer } from "./Styles";
+import { CreatePlaylist } from "./Styles";
+import { AllPlaylistisContainer } from "./Styles";
+import { PlaylistContainer } from "./Styles";
+import { GoToPlaylistContainer } from "./Styles";
+import { Delete } from "./Styles";
 
 export default class AllPlaylists extends React.Component {
     state = {
@@ -14,6 +22,8 @@ export default class AllPlaylists extends React.Component {
     componentDidMount(){
         this.getAllPlaylists();
     }
+
+    //Funções de evento
 
     handleInputPlaylist = (event) => {
         this.setState({
@@ -27,8 +37,11 @@ export default class AllPlaylists extends React.Component {
             console.log("enter");
         };
     };
+
+    //Requests
     
-    createPlaylist = async () => {
+    createPlaylist = async (e) => {
+        e.preventDefault();
         const body = {
             name: this.state.playListName
         };
@@ -38,6 +51,7 @@ export default class AllPlaylists extends React.Component {
             alert("Nova playlist criada!");
             this.setState({ playListName: "" });
             this.getAllPlaylists();
+            console.log(response)
         } catch (err) {
             alert(`Opa, não foi possível criar essa playlist. O erro foi:\n${err.response.data.message}`)
         };
@@ -54,14 +68,16 @@ export default class AllPlaylists extends React.Component {
     };
 
     deletePlaylist = async (id) => {
-        console.log(id)
-        try {
-            const response = await axios.delete(`${BASE_URL}/${id}`, axiosConfig);
-            console.log(response);
-            this.getAllPlaylists();
-        } catch (err) {
-            console.log("Deu erro", err.response.data)
-        }
+        const deleteConfirmation = "Tem certeza de que deseja deletar esta playlist?"
+        if (window.confirm(deleteConfirmation) === true) {
+            try {
+                const response = await axios.delete(`${BASE_URL}/${id}`, axiosConfig);
+                console.log(response);
+                this.getAllPlaylists();
+            } catch (err) {
+                console.log("Deu erro", err.response.data)
+            };
+        };
     };
 
     render() {
@@ -69,29 +85,37 @@ export default class AllPlaylists extends React.Component {
         
         const playlists = this.state.playlistsList.map( (list) => {
             return (
-                <div 
-                    key={list.id}
-                    onClick={() => this.props.goToAllSongsPage(list.id, list.name)}
-                >
-                    <img src={imgPlaylist} alt="Ícone de playlist de música" />
-                    {list.name}
-                    <button onClick={() => this.deletePlaylist(list.id)}>
+                <PlaylistContainer key={list.id}>
+                    <GoToPlaylistContainer 
+                        onClick={() => this.props.goToAllSongsPage(list.id, list.name)}
+                        title="Ver playlist"
+                    >
+                        <img src={imgCd} alt="Ícone de playlist de música" />
+                        <h3>{list.name}</h3>
+                    </GoToPlaylistContainer>
+                    <Delete onClick={() => this.deletePlaylist(list.id)} title="Deletar playlist">
                         <img src={imgDelete} alt="Ícone de lixeira" />
-                    </button>
-                </div>
+                    </Delete>
+                </PlaylistContainer>
             );
         });
         return (
-            <div>
-                <input
-                    placeholder="Nome da playlist"
-                    value={this.state.playListName}
-                    onChange={this.handleInputPlaylist}
-                    onKeyPress={this.inputPressEnter}
-                />
-                <button onClick={this.createPlaylist} >Criar playlist</button>
-                {playlists}
-            </div>
+            <MainContainer>
+                <h2>Suas playlists</h2>
+                <AllPlaylistisContainer>{playlists}</AllPlaylistisContainer>
+                <form>
+                    <CreatePlaylist>
+                        <legend>Nova Playlist</legend>
+                        <input
+                            placeholder="Nome da playlist"
+                            value={this.state.playListName}
+                            onChange={this.handleInputPlaylist}
+                            onKeyPress={this.inputPressEnter}
+                        />
+                        <button onClick={this.createPlaylist} >Criar</button>
+                    </CreatePlaylist>
+                </form>
+            </MainContainer>
         );
     };
 }
