+++ modulo2/projeto-labefy/labefy/src/pages/SelectedPlaylist/SelectedPlaylist.js
@@ -2,7 +2,15 @@ import React from "react";
 import axios from "axios";
 import { BASE_URL } from "../../constants/urls";
 import { axiosConfig } from "../../constants/headers";
-import imgDelete from "../AllPlaylists/img/delete.png"
+import imgDelete from "../../img/delete.png"
+
+import { MainContainer } from "./Styles";
+import { Playlist } from "./Styles";
+import { AllSongsContainer } from "./Styles";
+import { SongContainer } from "./Styles";
+import { AddSong } from "./Styles";
+import { Delete } from "./Styles";
+import { Song } from "./Styles";
 
 export default class SelectedPlaylist extends React.Component {
     state = {
@@ -18,6 +26,8 @@ export default class SelectedPlaylist extends React.Component {
         this.getPlaylistTracks(this.props.idPlaylist);
     }
 
+    // Funções de Evento
+
     handleInputName = (event) => {
         this.setState({ newSongName: event.target.value });
     };
@@ -30,9 +40,37 @@ export default class SelectedPlaylist extends React.Component {
         this.setState({ newUrl: event.target.value });
     };
 
-    onClickAddTrack = () => {
+    handleInputPressEnter = (event) => {
+        if (event.key === 'Enter') {
+            const form = event.target.form;
+            console.log("form", form);
+            const index = [...form].indexOf(event.target);
+            form.elements[index + 1].focus();
+            event.preventDefault();
+            console.log("enter");
+        };
+    }
+
+    lastInputPressEnter = (event) => {
+        if (event.key === 'Enter') {
+            this.onClickAddTrack();
+            console.log("enter");
+        };
+    };
+
+    onClickAddTrack = (e) => {
+        e.preventDefault();
         this.addTrackToPlaylist(this.props.idPlaylist);
     }
+
+    trackId = (url) => {
+        const slashSplit = url.split("/");
+        const questionMarkSplit = slashSplit[4].split("?")
+        const trackId = questionMarkSplit[0];
+        return `https://open.spotify.com/embed/track/${trackId}`
+    }
+
+    //Requests
 
     getPlaylistTracks = async (id) => {
         try {
@@ -59,6 +97,11 @@ export default class SelectedPlaylist extends React.Component {
             console.log(response);
             alert("Nova música foi adcionada à sua playlist!");
             this.getPlaylistTracks(this.props.idPlaylist);
+            this.setState({
+                newSongName: "",
+                newArtist: "",
+                newUrl: ""                
+            })
 
         } catch (err) {
             console.log(err.response)
@@ -67,12 +110,15 @@ export default class SelectedPlaylist extends React.Component {
     }
     
     removeTrackFromPlaylist = async (idPlaylist, idTrack) => {
-        try {
-            const response = await axios.delete(`${BASE_URL}/${idPlaylist}/tracks/${idTrack}`, axiosConfig);
-            console.log("Deu certo remover", response)
-            this.getPlaylistTracks(this.props.idPlaylist);
-        } catch (err) {
-            console.log("Deu erro", err.response.data)
+        const deleteConfirmation = "Tem certeza de que deseja deletar esta música?"
+        if (window.confirm(deleteConfirmation) === true) {
+            try {
+                const response = await axios.delete(`${BASE_URL}/${idPlaylist}/tracks/${idTrack}`, axiosConfig);
+                console.log("Deu certo remover", response)
+                this.getPlaylistTracks(this.props.idPlaylist);
+            } catch (err) {
+                console.log("Deu erro", err.response.data)
+            };
         };
     };
 
@@ -80,41 +126,70 @@ export default class SelectedPlaylist extends React.Component {
         console.log(this.state)
         const tracks = this.state.tracksList.map( (song) => {
             return (
-                <div key={song.id}>
-                    <p>Música: {song.name}</p>
-                    <p>Artista: {song.artist}</p>
-                    <p>Link: {song.url}</p>
-                    <button onClick={() => this.removeTrackFromPlaylist(this.props.idPlaylist, song.id)}>
+                <SongContainer key={song.id}>
+                    <Song>
+                        <iframe 
+                            src={this.trackId(song.url)} 
+                            width="80" height="80" 
+                            frameborder="0" 
+                            allowtransparency="true" 
+                            allow="encrypted-media" 
+                        />
+                        <div>
+                            <p><strong>{song.name}</strong></p>
+                            <p>by {song.artist}</p>
+                        </div>
+
+                    </Song>
+                    
+                    <Delete onClick={() => this.removeTrackFromPlaylist(this.props.idPlaylist, song.id)} title="Deletar música">
                         <img src={imgDelete}/>
-                    </button>
-                </div>
+                    </Delete>
+                </SongContainer>
             );
         });
         return (
-            <div>
-                <p><strong>{this.props.namePlaylist}</strong></p>
-                {this.state.quantity > 1 || this.state.quantity === 0
-                    ? <p>{this.state.quantity} faixas</p>
-                    : <p>{this.state.quantity} faixa</p>
-                }
-                <input 
-                    placeholder="Nome da música" 
-                    value={this.state.newSongName}
-                    onChange={this.handleInputName}
-                />
-                <input 
-                    placeholder="Nome do(a) artista ou da banda" 
-                    value={this.state.newArtist}
-                    onChange={this.handleInputArtist}
-                />
-                <input 
-                    placeholder="url da música" 
-                    value={this.state.newUrl}
-                    onChange={this.handleInputUrl}
-                />
-                <button  onClick={this.onClickAddTrack}>Adicionar música</button>
-                {tracks}
-            </div>
+            <MainContainer>
+                <h2>Suas músicas</h2>
+
+                <button onClick={this.props.goToPlaylistsPage}>Voltar para playlists</button>
+                
+                <Playlist>
+                    <h3><strong>{this.props.namePlaylist.toUpperCase()}</strong></h3>
+                    {this.state.quantity > 1 || this.state.quantity === 0
+                        ? <p>{this.state.quantity} faixas</p>
+                        : <p>{this.state.quantity} faixa</p>
+                    }
+                </Playlist>
+                
+                <AllSongsContainer>{tracks}</AllSongsContainer>
+
+                <form>
+                    <AddSong >
+                        <legend>Adicionar música</legend>
+                        <input 
+                            placeholder="Nome da música" 
+                            value={this.state.newSongName}
+                            onChange={this.handleInputName}
+                            onKeyPress={this.handleInputPressEnter}
+                        />
+                        <input 
+                            placeholder="Nome do(a) artista ou da banda" 
+                            value={this.state.newArtist}
+                            onChange={this.handleInputArtist}
+                            onKeyPress={this.handleInputPressEnter}
+                        />
+                        <input 
+                            placeholder="url da música" 
+                            value={this.state.newUrl}
+                            onChange={this.handleInputUrl}
+                            onKeyPress={this.lastInputPressEnter}
+                        />
+                        <button  onClick={this.onClickAddTrack}>Adicionar música</button>
+                    </AddSong>
+                </form>
+                
+            </MainContainer>
         );
     };
 };
